[IMP] l10n_it_edi_pa: merging in l10n_it_edi
Module `l10n_it_edi_pa` was created for Split Payment
because stored fields couldn't be added to an existing module.
We are amending this by merging it into module `l10n_it_edi`.

Task link: https://www.odoo.com/web#id=3513075&model=project.task
task-3513075
