+++ addons/l10n_it_edi/data/invoice_it_template.xml
@@ -227,10 +227,10 @@
 </template>
 
 <template id="account_invoice_FatturaPA_origin_document">
-                <IdDocumento t-if="origin_document_name" t-esc="format_alphanumeric(origin_document_name[:20])"/>
+                <IdDocumento t-if="origin_document_name" t-esc="format_alphanumeric(origin_document_name, 20)"/>
                 <Data t-if="origin_document_date" t-esc="format_date(origin_document_date)"/>
-                <CodiceCUP t-if="cup" t-esc="format_alphanumeric(cup)[-15:]"/>
-                <CodiceCIG t-if="cig" t-esc="format_alphanumeric(cig)[-15:]"/>
+                <CodiceCUP t-if="cup" t-esc="format_alphanumeric(cup, 15)"/>
+                <CodiceCIG t-if="cig" t-esc="format_alphanumeric(cig, 15)"/>
 </template>
 
     </data>
