+++ addons/l10n_it_edi/i18n/l10n_it_edi.pot
@@ -473,11 +473,6 @@ msgid "Italian EDI"
 msgstr ""
 
 #. module: l10n_it_edi
-#: model:ir.model,name:l10n_it_edi.model_account_move
-msgid "Journal Entry"
-msgstr ""
-
-#. module: l10n_it_edi
 #: model:ir.model.fields,field_description:l10n_it_edi.field_account_bank_statement_line__l10n_it_edi_attachment_file
 #: model:ir.model.fields,field_description:l10n_it_edi.field_account_move__l10n_it_edi_attachment_file
 #: model:ir.model.fields,field_description:l10n_it_edi.field_account_payment__l10n_it_edi_attachment_file
@@ -1402,4 +1397,123 @@ msgstr ""
 #: code:addons/l10n_it_edi/models/account_move.py:0
 #, python-format
 msgid "from XML file:"
+msgstr ""
+
+#. module: l10n_it_edi
+#: model_terms:ir.ui.view,arch_db:l10n_it_edi.report_invoice_document
+msgid "<b>CIG: </b>"
+msgstr ""
+
+#. module: l10n_it_edi
+#: model_terms:ir.ui.view,arch_db:l10n_it_edi.report_invoice_document
+msgid "<b>CUP: </b>"
+msgstr ""
+
+#. module: l10n_it_edi
+#: model_terms:ir.ui.view,arch_db:l10n_it_edi.report_invoice_document
+msgid "<b>Document Date: </b>"
+msgstr ""
+
+#. module: l10n_it_edi
+#: model:ir.model.fields.selection,name:l10n_it_edi.selection__account_move__l10n_it_origin_document_type__agreement
+msgid "Agreement"
+msgstr ""
+
+#. module: l10n_it_edi
+#: model:ir.model.fields,field_description:l10n_it_edi.field_account_bank_statement_line__l10n_it_cig
+#: model:ir.model.fields,field_description:l10n_it_edi.field_account_move__l10n_it_cig
+#: model:ir.model.fields,field_description:l10n_it_edi.field_account_payment__l10n_it_cig
+msgid "CIG"
+msgstr ""
+
+#. module: l10n_it_edi
+#: model:ir.model.fields,field_description:l10n_it_edi.field_account_bank_statement_line__l10n_it_cup
+#: model:ir.model.fields,field_description:l10n_it_edi.field_account_move__l10n_it_cup
+#: model:ir.model.fields,field_description:l10n_it_edi.field_account_payment__l10n_it_cup
+msgid "CUP"
+msgstr ""
+
+#. module: l10n_it_edi
+#: model:ir.model.fields.selection,name:l10n_it_edi.selection__account_move__l10n_it_origin_document_type__contract
+msgid "Contract"
+msgstr ""
+
+#. module: l10n_it_edi
+#: model:ir.model.fields,field_description:l10n_it_edi.field_account_bank_statement_line__l10n_it_partner_pa_index
+#: model:ir.model.fields,field_description:l10n_it_edi.field_account_move__l10n_it_partner_pa_index
+#: model:ir.model.fields,field_description:l10n_it_edi.field_account_payment__l10n_it_partner_pa_index
+msgid "Destination Code"
+msgstr ""
+
+#. module: l10n_it_edi
+#: model:ir.model,name:l10n_it_edi.model_account_move
+msgid "Journal Entry"
+msgstr ""
+
+#. module: l10n_it_edi
+#: model:ir.model.fields,help:l10n_it_edi.field_account_bank_statement_line__l10n_it_partner_pa_index
+#: model:ir.model.fields,help:l10n_it_edi.field_account_move__l10n_it_partner_pa_index
+#: model:ir.model.fields,help:l10n_it_edi.field_account_payment__l10n_it_partner_pa_index
+msgid ""
+"Must contain the 6-character (or 7) code, present in the PA              "
+"Index in the information relative to the electronic invoicing service,"
+"              associated with the office which, within the addressee "
+"administration, deals              with receiving (and processing) the "
+"invoice."
+msgstr ""
+
+#. module: l10n_it_edi
+#: model:ir.model.fields,field_description:l10n_it_edi.field_account_bank_statement_line__l10n_it_origin_document_date
+#: model:ir.model.fields,field_description:l10n_it_edi.field_account_move__l10n_it_origin_document_date
+#: model:ir.model.fields,field_description:l10n_it_edi.field_account_payment__l10n_it_origin_document_date
+msgid "Origin Document Date"
+msgstr ""
+
+#. module: l10n_it_edi
+#: model:ir.model.fields,field_description:l10n_it_edi.field_account_bank_statement_line__l10n_it_origin_document_name
+#: model:ir.model.fields,field_description:l10n_it_edi.field_account_move__l10n_it_origin_document_name
+#: model:ir.model.fields,field_description:l10n_it_edi.field_account_payment__l10n_it_origin_document_name
+msgid "Origin Document Name"
+msgstr ""
+
+#. module: l10n_it_edi
+#: model:ir.model.fields,field_description:l10n_it_edi.field_account_bank_statement_line__l10n_it_origin_document_type
+#: model:ir.model.fields,field_description:l10n_it_edi.field_account_move__l10n_it_origin_document_type
+#: model:ir.model.fields,field_description:l10n_it_edi.field_account_payment__l10n_it_origin_document_type
+msgid "Origin Document Type"
+msgstr ""
+
+#. module: l10n_it_edi
+#: model:ir.model.fields,help:l10n_it_edi.field_account_bank_statement_line__l10n_it_cup
+#: model:ir.model.fields,help:l10n_it_edi.field_account_move__l10n_it_cup
+#: model:ir.model.fields,help:l10n_it_edi.field_account_payment__l10n_it_cup
+msgid "Public Investment Unique Identifier"
+msgstr ""
+
+#. module: l10n_it_edi
+#: model:ir.model.fields.selection,name:l10n_it_edi.selection__account_move__l10n_it_origin_document_type__purchase_order
+msgid "Purchase Order"
+msgstr ""
+
+#. module: l10n_it_edi
+#: model:ir.model.fields,help:l10n_it_edi.field_account_bank_statement_line__l10n_it_cig
+#: model:ir.model.fields,help:l10n_it_edi.field_account_move__l10n_it_cig
+#: model:ir.model.fields,help:l10n_it_edi.field_account_payment__l10n_it_cig
+msgid "Tender Unique Identifier"
+msgstr ""
+
+#. module: l10n_it_edi
+#. odoo-python
+#: code:addons/l10n_it_edi/models/account_move.py:0
+#, python-format
+msgid "The Origin Document Date cannot be in the future."
+msgstr ""
+
+#. module: l10n_it_edi
+#. odoo-python
+#: code:addons/l10n_it_edi/models/account_move.py:0
+#, python-format
+msgid ""
+"This invoice targets the Public Administration, please fill out Origin "
+"Document Type field in the Electronic Invoicing tab."
 msgstr ""
