+++ addons/l10n_it_edi/tests/test_edi_pa.py
