--- addons/l10n_it_edi_pa/i18n/it.po
@@ -1,146 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-# 	* l10n_it_edi_pa
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 16.0+e\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2023-08-31 09:44+0000\n"
-"PO-Revision-Date: 2023-08-31 09:44+0000\n"
-"Last-Translator: \n"
-"Language-Team: \n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: \n"
-
-#. module: l10n_it_edi_pa
-#: model_terms:ir.ui.view,arch_db:l10n_it_edi_pa.report_invoice_document
-msgid "<b>CIG: </b>"
-msgstr ""
-
-#. module: l10n_it_edi_pa
-#: model_terms:ir.ui.view,arch_db:l10n_it_edi_pa.report_invoice_document
-msgid "<b>CUP: </b>"
-msgstr ""
-
-#. module: l10n_it_edi_pa
-#: model_terms:ir.ui.view,arch_db:l10n_it_edi_pa.report_invoice_document
-msgid "<b>Document Date: </b>"
-msgstr "<b>Data Documento: </b>"
-
-#. module: l10n_it_edi_pa
-#: model:ir.model.fields.selection,name:l10n_it_edi_pa.selection__account_move__l10n_it_origin_document_type__agreement
-msgid "Agreement"
-msgstr "Convenzione"
-
-#. module: l10n_it_edi_pa
-#: model:ir.model.fields,field_description:l10n_it_edi_pa.field_account_bank_statement_line__l10n_it_cig
-#: model:ir.model.fields,field_description:l10n_it_edi_pa.field_account_move__l10n_it_cig
-#: model:ir.model.fields,field_description:l10n_it_edi_pa.field_account_payment__l10n_it_cig
-msgid "CIG"
-msgstr ""
-
-#. module: l10n_it_edi_pa
-#: model:ir.model.fields,field_description:l10n_it_edi_pa.field_account_bank_statement_line__l10n_it_cup
-#: model:ir.model.fields,field_description:l10n_it_edi_pa.field_account_move__l10n_it_cup
-#: model:ir.model.fields,field_description:l10n_it_edi_pa.field_account_payment__l10n_it_cup
-msgid "CUP"
-msgstr ""
-
-#. module: l10n_it_edi_pa
-#: model:ir.model.fields.selection,name:l10n_it_edi_pa.selection__account_move__l10n_it_origin_document_type__contract
-msgid "Contract"
-msgstr "Contratto"
-
-#. module: l10n_it_edi_pa
-#: model:ir.model.fields,field_description:l10n_it_edi_pa.field_account_bank_statement_line__l10n_it_partner_pa_index
-#: model:ir.model.fields,field_description:l10n_it_edi_pa.field_account_move__l10n_it_partner_pa_index
-#: model:ir.model.fields,field_description:l10n_it_edi_pa.field_account_payment__l10n_it_partner_pa_index
-msgid "Destination Code"
-msgstr "Codice Destinatario"
-
-#. module: l10n_it_edi_pa
-#: model:ir.model,name:l10n_it_edi_pa.model_account_edi_format
-msgid "EDI format"
-msgstr "Formato EDI"
-
-#. module: l10n_it_edi_pa
-#: model:ir.model,name:l10n_it_edi_pa.model_account_move
-msgid "Journal Entry"
-msgstr "Movimento Contabile"
-
-#. module: l10n_it_edi_pa
-#: model:ir.model.fields,help:l10n_it_edi_pa.field_account_bank_statement_line__l10n_it_partner_pa_index
-#: model:ir.model.fields,help:l10n_it_edi_pa.field_account_move__l10n_it_partner_pa_index
-#: model:ir.model.fields,help:l10n_it_edi_pa.field_account_payment__l10n_it_partner_pa_index
-msgid ""
-"Must contain the 6-character (or 7) code, present in the PA              "
-"Index in the information relative to the electronic invoicing service,"
-"              associated with the office which, within the addressee "
-"administration, deals              with receiving (and processing) the "
-"invoice."
-msgstr ""
-"Deve contenere il codice da 6-7 caratteri presente nell'indice della PA "
-"nelle informazioni relative al servizio di Fatturazione Elettronica, "
-"associato all'ufficio dell'amministrazione che riceverá (ed elaborerá) la "
-"Fattura."
-
-#. module: l10n_it_edi_pa
-#: model:ir.model.fields,field_description:l10n_it_edi_pa.field_account_bank_statement_line__l10n_it_origin_document_date
-#: model:ir.model.fields,field_description:l10n_it_edi_pa.field_account_move__l10n_it_origin_document_date
-#: model:ir.model.fields,field_description:l10n_it_edi_pa.field_account_payment__l10n_it_origin_document_date
-msgid "Origin Document Date"
-msgstr "Data Documento Origine"
-
-#. module: l10n_it_edi_pa
-#: model:ir.model.fields,field_description:l10n_it_edi_pa.field_account_bank_statement_line__l10n_it_origin_document_name
-#: model:ir.model.fields,field_description:l10n_it_edi_pa.field_account_move__l10n_it_origin_document_name
-#: model:ir.model.fields,field_description:l10n_it_edi_pa.field_account_payment__l10n_it_origin_document_name
-msgid "Origin Document Name"
-msgstr "Nome Documento Origine"
-
-#. module: l10n_it_edi_pa
-#: model:ir.model.fields,field_description:l10n_it_edi_pa.field_account_bank_statement_line__l10n_it_origin_document_type
-#: model:ir.model.fields,field_description:l10n_it_edi_pa.field_account_move__l10n_it_origin_document_type
-#: model:ir.model.fields,field_description:l10n_it_edi_pa.field_account_payment__l10n_it_origin_document_type
-msgid "Origin Document Type"
-msgstr "Tipo Documento Origine"
-
-#. module: l10n_it_edi_pa
-#: model:ir.model.fields,help:l10n_it_edi_pa.field_account_bank_statement_line__l10n_it_cup
-#: model:ir.model.fields,help:l10n_it_edi_pa.field_account_move__l10n_it_cup
-#: model:ir.model.fields,help:l10n_it_edi_pa.field_account_payment__l10n_it_cup
-msgid "Public Investment Unique Identifier"
-msgstr "CUP"
-
-#. module: l10n_it_edi_pa
-#: model:ir.model.fields.selection,name:l10n_it_edi_pa.selection__account_move__l10n_it_origin_document_type__purchase_order
-msgid "Purchase Order"
-msgstr "Ordine d'Acquisto"
-
-#. module: l10n_it_edi_pa
-#: model:ir.model.fields,help:l10n_it_edi_pa.field_account_bank_statement_line__l10n_it_cig
-#: model:ir.model.fields,help:l10n_it_edi_pa.field_account_move__l10n_it_cig
-#: model:ir.model.fields,help:l10n_it_edi_pa.field_account_payment__l10n_it_cig
-msgid "Tender Unique Identifier"
-msgstr "CIG"
-
-#. module: l10n_it_edi_pa
-#. odoo-python
-#: code:addons/l10n_it_edi_pa/models/account_edi_format.py:0
-#, python-format
-msgid "The Origin Document Date cannot be in the future."
-msgstr "La Data Documento Origine non può essere nel futuro."
-
-#. module: l10n_it_edi_pa
-#. odoo-python
-#: code:addons/l10n_it_edi_pa/models/account_edi_format.py:0
-#, python-format
-msgid ""
-"This invoice targets the Public Administration, please fill out Origin "
-"Document Type field in the Electronic Invoicing tab."
-msgstr ""
-"Questa fattura ha come cliente la Pubblica Amministrazione, per favore "
-"riempi il Tipo Documento Origine nel tab Fatturazione Elettronica"
