--- addons/l10n_it_edi_pa/tests/__init__.py
@@ -1,3 +1,0 @@
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-
-from . import test_edi_pa
