--- addons/l10n_it_edi_pa/views/account_move_view.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-    <record id="account_invoice_form_l10n_it_pa" model="ir.ui.view">
-        <field name="name">account.move.form.l10n.it.pa</field>
-        <field name="model">account.move</field>
-        <field name="priority">20</field>
-        <field name="inherit_id" ref="l10n_it_edi.account_invoice_form_l10n_it"/>
-        <field name="arch" type="xml">
-        <data>
-            <xpath expr="//page[@name='electronic_invoicing']/group" position="inside">
-                <field name="l10n_it_partner_pa" invisible="1"/>
-                <group invisible="not l10n_it_partner_pa">
-                    <field name="l10n_it_origin_document_type" readonly="state != 'draft'"/>
-                    <field name="l10n_it_origin_document_name" readonly="state != 'draft'"/>
-                    <field name="l10n_it_origin_document_date" readonly="state != 'draft'"/>
-                    <field name="l10n_it_cig" readonly="state != 'draft'"/>
-                    <field name="l10n_it_cup" readonly="state != 'draft'"/>
-                </group>
-            </xpath>
-        </data>
-        </field>
-    </record>
-</odoo>
